changes path of pages
