+++ server/models/index.js
@@ -2,4 +2,4 @@ const User = require('./User');
 const Category = require('./Category');
 
 
-module.exports = { User, Product, Category, Order };
+module.exports = { User, Category, };
